Update tests for linux

+++ Tests/PSGetInstallModule.Tests.ps1
@@ -124,7 +124,7 @@ function SuiteSetup {
 }
 
 function SuiteCleanup {
-    if(Test-Path $script:moduleSourcesBackupFilePath)
+    if($script:moduleSourcesBackupFilePath -and (Test-Path $script:moduleSourcesBackupFilePath))
     {
         Move-Item $script:moduleSourcesBackupFilePath $script:moduleSourcesFilePath -Force
     }
@@ -175,6 +175,143 @@ Describe PowerShell.PSGet.InstallModuleTests -Tags 'BVT','InnerLoop' {
         PSGetTestUtils\Uninstall-Module ContosoClient
         PSGetTestUtils\Uninstall-Module DscTestModule
     }
+
+    
+    # Purpose: Install a module with current user scope parameter for non-admin User
+    #
+    # Action: Try to install a module with current user scope in a non-admin console
+    #
+    # Expected Result: It should succeed and install only to current user
+    #
+    It "InstallModuleWithCurrentUserScopeParameterForNonAdminUser" {
+        $PSprocess = "pwsh"
+        if ($script:IsWindows) {
+            $PSprocess = "PowerShell.exe";
+        }
+
+        $NonAdminConsoleOutput = Join-Path ([System.IO.Path]::GetTempPath()) 'nonadminconsole-out.txt'
+
+        Start-Process $PSprocess -ArgumentList '$null = Install-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force -Scope CurrentUser;
+                                                              $null = Import-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force;
+                                                              if(-not (Get-PSRepository -Name INTGallery -ErrorAction SilentlyContinue)) {
+                                                                Register-PSRepository -Name INTGallery -SourceLocation https://dtlgalleryint.cloudapp.net/api/v2/ -InstallationPolicy Trusted
+                                                              }
+                                                              Install-Module -Name ContosoServer -Repository INTGallery -Scope CurrentUser;
+                                                              Get-InstalledModule ContosoServer | Format-List Name, InstalledLocation' `
+                                               -Credential $script:credential `
+                                               -Wait `
+                                               -RedirectStandardOutput $NonAdminConsoleOutput
+
+        waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Module on non-admin console failed to complete"
+        $content = Get-Content $NonAdminConsoleOutput
+        RemoveItem $NonAdminConsoleOutput
+
+        AssertNotNull ($content) "Install module with CurrentUser scope on non-admin user console should succeed"
+        Assert ($content -match "ContosoServer") "Module did not install correctly"
+        if ($script:IsWindows) {
+            Assert ($content -match "Documents") "Module did not install to the correct location"
+        }
+        else {
+            Assert ($content -match "home") "Module did not install to the correct location"
+        }
+    } `
+    -Skip:$(
+        $whoamiValue = (whoami)
+
+        ($whoamiValue -eq "NT AUTHORITY\SYSTEM") -or
+        ($whoamiValue -eq "NT AUTHORITY\LOCAL SERVICE") -or
+        ($whoamiValue -eq "NT AUTHORITY\NETWORK SERVICE") -or
+        ($PSVersionTable.PSVersion -lt '4.0.0') -or
+        # Temporarily skip tests until .NET Core is updated to v2.1
+        ($PSEdition -eq 'Core')
+    )
+
+    # Purpose: Install a module with all users scope parameter for non-admin user
+    #
+    # Action: Try to install a module with all users scope in a non-admin console
+    #
+    # Expected Result: It should fail with an error
+    #
+    It "InstallModuleWithAllUsersScopeParameterForNonAdminUser" {
+        $PSprocess = "pwsh"
+        if ($script:IsWindows) {
+            $PSprocess = "PowerShell.exe";
+        }
+
+        $NonAdminConsoleOutput = Join-Path ([System.IO.Path]::GetTempPath()) 'nonadminconsole-out.txt'
+
+        Start-Process $PSprocess -ArgumentList '$null = Install-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force -Scope AllUsers;
+                                                              $null = Import-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force;
+                                                              Install-Module -Name ContosoServer -scope AllUsers -ErrorAction SilentlyContinue -Repository INTGallery
+                                                              Get-InstalledModule -Name ContosoServer | Format-List Name, InstalledLocation' `
+                                               -Credential $script:credential `
+                                               -Wait `
+                                               -RedirectStandardOutput $NonAdminConsoleOutput
+
+
+        waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Module on non-admin console failed to complete"
+        $content = Get-Content $NonAdminConsoleOutput
+        RemoveItem $NonAdminConsoleOutput
+
+        AssertNotNull ($content) "Install module with CurrentUser scope on non-admin user console should not succeed"
+        Assert ($content -match "Administrator rights are required to install packages") "Install module with AllUsers scope on non-admin user console should fail, $content"
+    } `
+    -Skip:$(
+        $whoamiValue = (whoami)
+
+        ($whoamiValue -eq "NT AUTHORITY\SYSTEM") -or
+        ($whoamiValue -eq "NT AUTHORITY\LOCAL SERVICE") -or
+        ($whoamiValue -eq "NT AUTHORITY\NETWORK SERVICE") -or
+        ($PSVersionTable.PSVersion -lt '4.0.0') -or
+        # Temporarily skip tests until .NET Core is updated to v2.1
+        ($PSEdition -eq 'Core')
+    )
+
+    # Purpose: Install a module with default scope parameter for non-admin user
+    #
+    # Action: Try to install a module with default (current user) scope in a non-admin console
+    #
+    # Expected Result: It should succeed and install only to current user
+    #
+    It "InstallModuleWithDefaultScopeParameterForNonAdminUser" {
+        $PSprocess = "pwsh"
+        if ($script:IsWindows) {
+            $PSprocess = "PowerShell.exe";
+        }
+
+        $NonAdminConsoleOutput = Join-Path ([System.IO.Path]::GetTempPath()) 'nonadminconsole-out.txt'
+
+        Start-Process $PSprocess -ArgumentList '$null = Install-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force -Scope CurrentUser;
+                                                              $null = Import-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force;
+                                                              Install-Module -Name ContosoServer -Repository INTGallery
+                                                              Get-InstalledModule -Name ContosoServer | Format-List Name, InstalledLocation' `
+                                               -Credential $script:credential `
+                                               -Wait `
+                                               -RedirectStandardOutput $NonAdminConsoleOutput
+
+        waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Module on non-admin console failed to complete"
+        $content = Get-Content $NonAdminConsoleOutput
+
+        RemoveItem $NonAdminConsoleOutput
+        AssertNotNull ($content) "Install module with CurrentUser scope on non-admin user console should succeed"
+        Assert ($content -match "ContosoServer") "Module did not install correctly"
+        if ($script:IsWindows) {
+            Assert ($content -match "Documents") "Module did not install to the correct location"
+        }
+        else {
+            Assert ($content -match "home") "Module did not install to the correct location"
+        }
+    } `
+    -Skip:$(
+        $whoamiValue = (whoami)
+
+        ($whoamiValue -eq "NT AUTHORITY\SYSTEM") -or
+        ($whoamiValue -eq "NT AUTHORITY\LOCAL SERVICE") -or
+        ($whoamiValue -eq "NT AUTHORITY\NETWORK SERVICE") -or
+        ($PSVersionTable.PSVersion -lt '4.0.0') -or
+        # Temporarily skip tests until .NET Core is updated to v2.1
+        ($PSEdition -eq 'Core')
+    )
 
     # Purpose: InstallNotAvailableModuleWithWildCard
     #
@@ -923,145 +1060,6 @@ Describe PowerShell.PSGet.InstallModuleTests -Tags 'BVT','InnerLoop' {
         }
     } `
     -Skip:$((-not (Get-Module PKI -ListAvailable)) -or ([Environment]::OSVersion.Version -lt '10.0'))
-
-    # Purpose: Install a module with current user scope parameter for non-admin User
-    #
-    # Action: Try to install a module with current user scope in a non-admin console
-    #
-    # Expected Result: It should succeed and install only to current user
-    #
-    It "InstallModuleWithCurrentUserScopeParameterForNonAdminUser" {
-        $PSprocess = "pwsh"
-        if ($script:IsWindows) {
-            $PSprocess = "PowerShell.exe";
-        }
-
-        $NonAdminConsoleOutput = Join-Path ([System.IO.Path]::GetTempPath()) 'nonadminconsole-out.txt'
-
-        Start-Process $PSprocess -ArgumentList 'Get-InstalledModule -Name ContosoServer | Uninstall-Module;
-                                                             $null = Install-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force -Scope CurrentUser;
-                                                              $null = Import-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force;
-                                                              if(-not (Get-PSRepository -Name INTGallery -ErrorAction SilentlyContinue)) {
-                                                                Register-PSRepository -Name INTGallery -SourceLocation https://dtlgalleryint.cloudapp.net/api/v2/ -InstallationPolicy Trusted
-                                                              }
-                                                              Install-Module -Name ContosoServer -Repository INTGallery -Scope CurrentUser;
-                                                              Get-InstalledModule ContosoServer | Format-List Name, InstalledLocation' `
-                                               -Credential $script:credential `
-                                               -Wait `
-                                               -RedirectStandardOutput $NonAdminConsoleOutput
-
-        waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Module on non-admin console failed to complete"
-        $content = Get-Content $NonAdminConsoleOutput
-        RemoveItem $NonAdminConsoleOutput
-
-        AssertNotNull ($content) "Install module with CurrentUser scope on non-admin user console should succeed"
-        Assert ($content -match "ContosoServer") "Module did not install correctly"
-        if ($script:IsWindows) {
-            Assert ($content -match "Documents") "Module did not install to the correct location"
-        }
-        else {
-            Assert ($content -match "home") "Module did not install to the correct location"
-        }
-    } `
-    -Skip:$(
-        $whoamiValue = (whoami)
-
-        ($whoamiValue -eq "NT AUTHORITY\SYSTEM") -or
-        ($whoamiValue -eq "NT AUTHORITY\LOCAL SERVICE") -or
-        ($whoamiValue -eq "NT AUTHORITY\NETWORK SERVICE") -or
-        ($PSVersionTable.PSVersion -lt '4.0.0') -or
-        # Temporarily skip tests until .NET Core is updated to v2.1
-        ($PSEdition -eq 'Core')
-    )
-
-    # Purpose: Install a module with all users scope parameter for non-admin user
-    #
-    # Action: Try to install a module with all users scope in a non-admin console
-    #
-    # Expected Result: It should fail with an error
-    #
-    It "InstallModuleWithAllUsersScopeParameterForNonAdminUser" {
-        $PSprocess = "pwsh"
-        if ($script:IsWindows) {
-            $PSprocess = "PowerShell.exe";
-        }
-
-        $NonAdminConsoleOutput = Join-Path ([System.IO.Path]::GetTempPath()) 'nonadminconsole-out.txt'
-
-        Start-Process $PSprocess -ArgumentList 'Get-InstalledModule -Name ContosoServer | Uninstall-Module;
-                                                              $null = Install-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force;
-                                                              $null = Import-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force;
-                                                              Install-Module -Name ContosoServer -scope AllUsers -ErrorAction SilentlyContinue -Repository INTGallery
-                                                              Get-InstalledModule -Name ContosoServer | Format-List Name, InstalledLocation' `
-                                               -Credential $script:credential `
-                                               -Wait `
-                                               -RedirectStandardOutput $NonAdminConsoleOutput
-
-
-        waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Module on non-admin console failed to complete"
-        $content = Get-Content $NonAdminConsoleOutput
-        RemoveItem $NonAdminConsoleOutput
-
-        AssertNotNull ($content) "Install module with CurrentUser scope on non-admin user console should not succeed"
-        Assert ($content -match "Administrator rights are required to install packages") "Install module with AllUsers scope on non-admin user console should fail, $content"
-    } `
-    -Skip:$(
-        $whoamiValue = (whoami)
-
-        ($whoamiValue -eq "NT AUTHORITY\SYSTEM") -or
-        ($whoamiValue -eq "NT AUTHORITY\LOCAL SERVICE") -or
-        ($whoamiValue -eq "NT AUTHORITY\NETWORK SERVICE") -or
-        ($PSVersionTable.PSVersion -lt '4.0.0') -or
-        # Temporarily skip tests until .NET Core is updated to v2.1
-        ($PSEdition -eq 'Core')
-    )
-
-    # Purpose: Install a module with default scope parameter for non-admin user
-    #
-    # Action: Try to install a module with default (current user) scope in a non-admin console
-    #
-    # Expected Result: It should succeed and install only to current user
-    #
-    It "InstallModuleWithDefaultScopeParameterForNonAdminUser" {
-        $PSprocess = "pwsh"
-        if ($script:IsWindows) {
-            $PSprocess = "PowerShell.exe";
-        }
-
-        $NonAdminConsoleOutput = Join-Path ([System.IO.Path]::GetTempPath()) 'nonadminconsole-out.txt'
-
-        Start-Process $PSprocess -ArgumentList 'Get-InstalledModule -Name ContosoServer | Uninstall-Module;
-                                                              $null = Install-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force -Scope CurrentUser;
-                                                              $null = Import-PackageProvider -Name NuGet -MinimumVersion 2.8.5.201 -Force;
-                                                              Install-Module -Name ContosoServer -Repository INTGallery
-                                                              Get-InstalledModule -Name ContosoServer | Format-List Name, InstalledLocation' `
-                                               -Credential $script:credential `
-                                               -Wait `
-                                               -RedirectStandardOutput $NonAdminConsoleOutput
-
-        waitFor {Test-Path $NonAdminConsoleOutput} -timeoutInMilliseconds $script:assertTimeOutms -exceptionMessage "Install-Module on non-admin console failed to complete"
-        $content = Get-Content $NonAdminConsoleOutput
-
-        RemoveItem $NonAdminConsoleOutput
-        AssertNotNull ($content) "Install module with CurrentUser scope on non-admin user console should succeed"
-        Assert ($content -match "ContosoServer") "Module did not install correctly"
-        if ($script:IsWindows) {
-            Assert ($content -match "Documents") "Module did not install to the correct location"
-        }
-        else {
-            Assert ($content -match "home") "Module did not install to the correct location"
-        }
-    } `
-    -Skip:$(
-        $whoamiValue = (whoami)
-
-        ($whoamiValue -eq "NT AUTHORITY\SYSTEM") -or
-        ($whoamiValue -eq "NT AUTHORITY\LOCAL SERVICE") -or
-        ($whoamiValue -eq "NT AUTHORITY\NETWORK SERVICE") -or
-        ($PSVersionTable.PSVersion -lt '4.0.0') -or
-        # Temporarily skip tests until .NET Core is updated to v2.1
-        ($PSEdition -eq 'Core')
-    )
 
     # Purpose: Install a module with current user scope parameter for admin user
     #
